requirements.txt
logging
click
loguru
numpy
pandas
plotly
kaleido
sqlalchemy
python-binance
typing
python-telegram-bot